Move hook file
